+++ Cpp/Executables/Regression/ERR_old.cpp
@@ -5,8 +5,8 @@
 #include <ERR_Regressor.hpp>
 #include <iostream>
 int main() {
-  size_t N_sims = 100; // 10000;
-  size_t N_pop = 60;
+  uint16_t N_sims = 100; // 10000;
+  uint16_t N_pop = 60;
   double p_ER = 1.0;
   using namespace FROLS;
   std::vector<std::string> df_names(N_sims);
@@ -23,11 +23,11 @@ int main() {
 
   std::cout << X << std::endl;
 
-  size_t d_max = 1;
-  size_t N_output_features = 16;
+  uint16_t d_max = 1;
+  uint16_t N_output_features = 16;
   using namespace FROLS::Features;
-  size_t Nx = X.cols();
-  size_t Nu = U.cols();
+  uint16_t Nx = X.cols();
+  uint16_t Nu = U.cols();
 
   FROLS::Features::Polynomial_Model model(Nx, Nu, N_output_features, d_max);
   double ERR_tol = 1e-1;
@@ -37,7 +37,7 @@ int main() {
 
   Vec x0 = X.row(0);
   double u0 = U(0, 0);
-  size_t Nt = 30;
+  uint16_t Nt = 30;
   Vec u = Vec::Ones(Nt) * u0;
   // print x0, u
   std::cout << "x0 = " << x0.transpose() << std::endl;
