+++ Cpp/include/FROLS_Path_Config.hpp
@@ -9,8 +9,8 @@ namespace FROLS {
     extern const char *FROLS_DATA_DIR;
     extern const char *FROLS_LOG_DIR;
 
-    std::string MC_filename(size_t N_pop, double p_ER, size_t iter, std::string network_type);
+    std::string MC_filename(uint16_t N_pop, double p_ER, uint16_t iter, std::string network_type);
 
-    std::string quantile_filename(size_t N_pop, double p_ER, size_t iter, std::string network_type);
+    std::string quantile_filename(uint16_t N_pop, double p_ER, uint16_t iter, std::string network_type);
 } //FROLS
 #endif
