+++ Cpp/static/DataFrame/DataFrame.cpp
@@ -19,7 +19,7 @@ namespace FROLS {
 
 
     void DataFrame::assign(const std::string &col_name,
-                           const std::vector<size_t> &col_data) {
+                           const std::vector<uint16_t> &col_data) {
 
         data[col_name] = std::make_shared<std::vector<double>>(col_data.size());
         for (int i = 0; i < col_data.size(); i++) {
@@ -33,7 +33,7 @@ namespace FROLS {
         data[col_name] = std::make_shared<std::vector<double>>(col_data);
     }
 
-    void DataFrame::assign(const std::string &col_name, size_t row, double value) {
+    void DataFrame::assign(const std::string &col_name, uint16_t row, double value) {
         if (data[col_name]->size() < row+1)
         {
             data[col_name]->resize(row+1);
@@ -50,8 +50,8 @@ namespace FROLS {
         return res;
     }
     void DataFrame::assign(const std::vector<std::string> &col_names,
-                           const std::vector<std::vector<size_t>> &col_data) {
-        for (size_t i = 0; i < col_names.size(); i++) {
+                           const std::vector<std::vector<uint16_t>> &col_data) {
+        for (uint16_t i = 0; i < col_names.size(); i++) {
             std::vector<double> col(col_data[i].size());
             std::copy(col_data[i].begin(), col_data[i].end(), col.begin());
             data[col_names[i]] = std::make_shared<std::vector<double>>(col);
@@ -60,7 +60,7 @@ namespace FROLS {
 
     void DataFrame::assign(const std::vector<std::string> &col_names,
                            const std::vector<std::vector<double>> &col_data) {
-        for (size_t i = 0; i < col_names.size(); i++) {
+        for (uint16_t i = 0; i < col_names.size(); i++) {
             *data[col_names[i]] = col_data[i];
         }
     }
@@ -88,12 +88,12 @@ namespace FROLS {
                     std::make_shared<std::vector<double>>(N_rows);
             data.insert({col_name, ptr});
         }
-        size_t row = 0;
+        uint16_t row = 0;
         N_rows = 0;
         while (std::getline(file, line)) {
             std::stringstream ss(line);
             std::string value;
-            size_t col = 0;
+            uint16_t col = 0;
             while (std::getline(ss, value, delimiter[0])) {
                 this->assign(col_names[col], row, std::stod(value));
                 col++;
@@ -118,8 +118,8 @@ namespace FROLS {
 
         resize();
         std::ofstream file(filename);
-        size_t N_cols = data.size();
-        size_t iter = 0;
+        uint16_t N_cols = data.size();
+        uint16_t iter = 0;
         for (auto &col: data) {
             iter++;
             file << col.first;
@@ -128,7 +128,7 @@ namespace FROLS {
             }
         }
         file << "\n";
-        for (size_t i = 0; i < N_rows; i++) {
+        for (uint16_t i = 0; i < N_rows; i++) {
             std::vector<double> coldata(data.size());
             std::transform(data.begin(), data.end(), coldata.begin(),
                            [&i](auto &col) { return col.second->operator[](i); });
@@ -141,7 +141,7 @@ namespace FROLS {
         }
     }
 
-    void DataFrame::resize(size_t N_rows) {
+    void DataFrame::resize(uint16_t N_rows) {
         if (!data.empty()) {
             for (auto &col: data) {
                 col.second->resize(N_rows);
@@ -151,20 +151,20 @@ namespace FROLS {
     }
 
     void DataFrame::resize() {
-        size_t rows = 0;
+        uint16_t rows = 0;
         for (const auto &col: data) {
-            rows = std::max({rows, col.second->size()});
+            rows = std::max({(size_t)rows, col.second->size()});
         }
         N_rows = rows;
     }
 
 
-    const std::vector<double> DataFrameStack::get_col(size_t frame, const std::string &col_name) {
+    const std::vector<double> DataFrameStack::get_col(uint16_t frame, const std::string &col_name) {
         assert(frame < dataframes.size());
         return *(dataframes[frame][col_name]);
     }
 
-    std::vector<double> DataFrameStack::get_row(size_t frame, size_t row) {
+    std::vector<double> DataFrameStack::get_row(uint16_t frame, uint16_t row) {
         std::vector<double> row_data;
         for (auto &col: dataframes[frame].data) {
             row_data.push_back(col.second->operator[](row));
@@ -172,8 +172,8 @@ namespace FROLS {
         return row_data;
     }
 
-    std::vector<size_t> DataFrameStack::get_N_rows() {
-        std::vector<size_t> N_rows(dataframes.size());
+    std::vector<uint16_t> DataFrameStack::get_N_rows() {
+        std::vector<uint16_t> N_rows(dataframes.size());
         std::generate(N_rows.begin(), N_rows.end(), [&,n = 0]() mutable { return dataframes[n].get_N_rows(); });
         return N_rows;
     }
